requirements.txt
ago==0.0.93
argcomplete==1.11.1
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
Automat==20.2.0
awscli==1.18.79
beautifulsoup4==4.8.0
botocore==1.17.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.5
docutils==0.15.2
docx2txt==0.8
dotmap==1.3.17
EbookLib==0.17.1
elasticsearch==7.7.1
extract-msg==0.23.1
feedfinder2==0.0.4
feedparser==5.2.1
gcloud==0.18.3
geojson==2.5.0
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-speech==1.3.2
google-cloud-texttospeech==2.0.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gTTS==2.1.1
gTTS-token==1.1.3
hjson==3.0.1
httplib2==0.17.0
hurry.filesize==0.9
hyperlink==19.0.0
idna==2.9
image==1.5.30
IMAPClient==2.1.0
incremental==17.5.0
isort==4.3.21
jieba3k==0.35.1
jmespath==0.10.0
joblib==0.14.1
langdetect==1.0.8
lazy-object-proxy==1.4.3
libcst==0.3.6
lxml==4.5.0
mccabe==0.6.1
mypy-extensions==0.4.3
news-please==1.5.3
newspaper3k==0.2.8
nltk==3.5
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
pafy==0.5.5
parsel==1.6.0
pdf2image==1.12.1
pdfminer.six==20181108
Pillow==7.1.1
plac==1.2.0
Protego==0.1.16
proto-plus==0.4.0
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pycryptodome==3.9.7
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyowm==3.0.0
PyPDF2==1.26.0
PySocks==1.7.1
pytesseract==0.3.3
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
python-vlc==3.0.7110
pytz==2019.3
pywin32==228
PyYAML==5.3.1
queuelib==1.5.0
readability-lxml==0.7.1
regex==2020.4.4
requests==2.23.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
Scrapy==2.1.0
service-identity==18.1.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==2.0
SpeechRecognition==3.8.1
sqlparse==0.3.1
tesseract==0.1.3
textract==1.6.3
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.45.0
Twisted==20.3.0
typing-extensions==3.7.4.2
typing-inspect==0.6.0
tzlocal==1.5.1
uritemplate==3.0.1
urllib3==1.25.8
w3lib==1.22.0
warcio==1.7.3
wget==3.2
wikipedia==1.4.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.8
youtube-dl==2020.6.6
zope.interface==5.1.0
